Bonle Version 1.0a
I have done a bunch of work on Bonle this past week as well as a little in winter break. It went from a tiny little demo that I put together in a day for Programming 1 to a fully fledged application. Some of the things I have done: I updated the human skeleton model to a better-looking and more anatomically correct version, I updated the UI slightly, its not good but its better than the Prototype Demo, I added both a Cat and a Trex to go along with the Human gamemodes, as well as a selection screen you are directed to on pressing play that allows you to choose which you want to play, and most importantly I updated the coloring algorithm to be better by graphing each of the bones to a node system and using Dijkstra's Algorithm on it, so now instead of an arbitrary spatial measurement(because we don't know what position the skeleton is in) the system is not dependent on the connections between bones. Selection of difficulties by the user will come in the next version, and either the next version or the subsequent version after that there will be a Practice game mode that includes a tutorial so you can learn how the game works.