Week of Work 2/20/2023
Been working a bunch this week so I haven't been doing regular commits, I had a four day weekend this week so I was able to get a bunch done including but not limited to getting the title page animations done, finishing all difficulties for creatures, designing a much better selection screen that includes the difficulties and the model selection on one screen, and fixing a ton of bugs. I started timing myself and it works out great for keeping yourself on task and I have found that when I do it its much easier to get things done. All thats left is to tie up some loose ends in terms of design, finish the settings/credits, and most importantly use the graphs that I have finished to make the practice mode. It should be done this coming week because I will be at my moms until next wednesday(today is monday)