Days of Work 2/21-22/2023
I finished implementing the new selection screen that eliminates the unnecessary scene for selecting difficulty. I implemented a quick example for the practice in the easy human because I need to present it but I still need to find a good solution for combining the meshes that doesn't require me to reimport every mesh in every difficulty.