Week of Work 2/27/2023
I am super close to done with the whole project, all I have left to do is make the specific skeletons for each of the practice modes other than the cat cause I finished that one, and maybe if I need to add credits I will add a credits page. Other than that its just little touch ups like the game over screen fades in and tells you how many attempts you did it in. Oh also the other major thing I have to do for the buisness club competition is make a leaderboard. Will be finished in the coming days because the buisness comp is on the 7th and the TSA comp is on the 15.